Refactored controls
